Implement authorization using Gates

+++ app/Http/Controllers/QuestionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionController extends Controller
 {
@@ -72,7 +73,11 @@ class QuestionController extends Controller
     public function edit(Question $question)
     {
         //
-		  return view("questions.edit", compact("question"));
+		if(Gate::allows('update-question', $question)){
+			return view("questions.edit", compact("question"));	
+		}
+		abort(403, 'Access denied');
+		  
     }
 
     /**
@@ -85,6 +90,9 @@ class QuestionController extends Controller
     public function update(AskQuestionRequest $request, Question $question)
     {
         //
+		if(Gate::denies('update-question', $question)){
+			abort(403, 'Access denied');
+		}
 		$question->update($request->only('title'));
 		return redirect('/questions')->with('success', 'Your question has been updated');
     }
@@ -98,6 +106,9 @@ class QuestionController extends Controller
     public function destroy(Question $question)
     {
         //
+		if(Gate::denies('delete-question', $question)){
+			abort(403, 'Access denied');	
+		}
 		$question->delete();
 		return redirect('/questions')->with('success', 'Your question has been deleted.');
     }
